requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.5
db-dtypes==1.1.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.1.0
fastjsonschema==2.16.2
fonttools==4.25.0
gensim==4.3.1
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
MarkupSafe==2.1.1
mistune==2.0.4
mkl-fft==1.3.6
mkl-service==2.4.0
munkres==1.1.4
nbclient==0.6.8
nbconvert==7.1.0
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==2.8.8
node2vec==0.4.6
notebook==6.4.12
numpy==1.25.0
packaging==21.3
panda==0.3.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.31
proto-plus==1.22.3
protobuf==4.24.1
psutil==5.9.2
pulearn==0.0.7
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
pyparsing==3.0.9
PyQt5==5.15.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pywin32==304
pywinpty==2.0.8
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.31.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
terminado==0.16.0
threadpoolctl==3.2.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.66.1
traitlets==5.4.0
urllib3==1.26.16
wcwidth==0.2.5
webencodings==0.5.1
wget==3.2
widgetsnbextension==4.0.3
